+++ src/routes/tickets.routes.js
@@ -1,0 +1,32 @@
+const express = require('express');
+const router = express.Router();
+const {
+  getTickets,
+  getTicketsAdmin,
+  getTicketsHelpDesk,
+  getOpenTicketsEmail,
+  getTicketsAll,
+  getTicket,
+  create_ticket,
+  update_ticket,
+  status_ticket,
+  solution_category
+} = require('../controllers/tickets.controller');
+
+// Rutas de tickets
+
+router.get('/tickets/:company_id/:fecha_inicial/:fecha_final', getTickets);
+router.get('/ticketsAdmin/:fecha_inicial/:fecha_final', getTicketsAdmin);
+router.get('/ticketsHelpDesk/:person_id/:fecha_inicial/:fecha_final', getTicketsHelpDesk);
+router.get('/open-tickets/:user_email', getOpenTicketsEmail);
+router.post('/tickets/create_ticket', create_ticket);
+router.put('/update_ticket/:ticket_id', update_ticket);
+router.get('/tickets/status_ticket', status_ticket);
+router.get('/tickets/solution_category', solution_category);
+router.get('/v1/tickets/:ticket_id', getTicket);
+
+
+router.get('/tickets/:company_id', getTicketsAll);
+
+
+module.exports = router;
